Schema generation refactoring. I was going to document Kinetic::Build::Schema::DB, but then realized that some of its documentation of naming conventions actually belonged in the Meta schema classes, where the names of things are returned by methods. But then I realized that the names of some things weren't returned there--namely primary key and foreign key constraints. So I created the methods for them, documented them, and then modified the schema generation classes to use them.
And as I was doing this, I realized that the Pg and SQLite schema generators
were using two different formats for naming foreign keys applied to primary
keys (for inheritance). So I made them consistent, changing them to include
both the key name of the parent and of the current class, so that they are
always unique, since SQLite requires unique trigger names.

I just realized that I forgot the tests for these new methods. I'll add those
momentarily.