The search class now has extra data validation for BETWEEN and ANY searches. It generates reasonable error messages if people try non-sensical searches and it also correctly handles dispatching for incomplete dates.
Meanwhile, the boundary condition issues that I encountered yesterday were
eliminated today by using a substr and concatenation technique that David
suggested.  This has also simplified much of the code in the SQLite.pm
store subclass.  This has also caused me to develop extra code in the
Incomplete.pm class to testing if two incomplete dates have the same
segments and generating "sortable strings" for the data which then get
compared to the concatenated date segments.

Incomplete date searches with BETWEEN and ANY are now implemented.