Initial work with proper Postgresql store tests.  PG does not store dates the way SQLite does, (it omits the "T" in date format), so the tests are going to manually convert things to DateTime objects as a stopgap measure until David has K::M accessors inflate objects.
There was a problem with the pg_teardown not properly recognizing that resources had been released, so some
code was added to for teardown to hang until the resources in PG were available.  Also have the PG.pm store
tests disconnecting from the database on teardown.

Many of the PG.pm tests have a "return" as the first statement in order to skip tests failures.  I will be 
pulling those out one-by-one until all tests pass.  At that point, the plan is to push those tests that are
duplicates back into a base class.

Also, there is some code to supress "subroutine redefined" warnings.  This is temporary until the refactoring
occurs.


The last_insert_id() routine for Postgres was not working.  We need to figure out why.