PG store tests now implemented.  All pass except the full text search (which will be implemented at a later date).
In the pg class, we have dates being cast as their respective datetime objects.  At some point
in the future, this code should be removed when we have the Kinetic::Meta tests inflating 
objects.

As it turns out, Postgres does not seem to recognize abs(), the pg methods now explicitly cast
the result of date substrings rather than let this be a side-effect of abs().  Also, most date
handling methods have been pushed back to db.pm.  Another problem is the DBD::SQLite allows
references in bind params and merely stringifies them.  DBD::Pg does not allow this, so we have
had to put a wrapper around the $sth->execute method to ensure that PG bind params properly
stringify if called if the params are objects.

I was also going to have some of the common tests refactored in this commit, but I was getting
very weird build failures that I could not diagnose.  As a result, I backed out that
refactoring to ensure that I could get something committed.