This is the version of the refactored code which is causing build problems.  For some reason, in TEST::Class::Kinetic::Build::Store::DB::Pg, as run through the "build.t" test, we fail the ok !$kbs->_db_exists($kbs->test_db_name) test and later hangs.
However, if I explicitly tear down the pg version of the database beforehand, the
build.t tests pass, so again, there is an issue with the teardown.  I can keep
working on the refactoring, though, as now that I have diagnosed the issue, it is
only a minor annoyance.

There is one test now failing in store.t, but I do not have time to fix it tonight.
I will take care of it tomorrow.