Refactored date handling dispatch to ensure that adding support for date handling in a new store is relatively easy (well, easier).  Many custom date handling methods have gone away now.
Incorporated inflation code into the test and pulled it out of the store class.
Much nicer now :)  This did require a bit of work in the tests, but it was not
too bad.

Updated pod_coverage tests to no longer skip modules with dynamically generated
subroutines.

Added new "defined_store_fields" to DateTime::Incomplete to allow easy testing
of which fields are useful.

Pushed all search operator munging back into the search class.  I am not sure if 
that was the best way to go.

Oh, almost forgot.  Both Pg and SQLite stores now handle dates themselves and 
the DB.pm class dies if the date handling methods are not implemented.