Initial XML serialization/deserialization code.  I am hard-pressed to explain why.  There is some duplicate database setup code in the XML and Store tests. I am unsure of the best way to refactor this.  Also, I had to write special code for the object ids.  This led to a couple of subtle bugs, but they were quickly worked out.
_force_inflation is such a useful routine which is now used in multiple tests
that I pushed this back to TEST::Class::Kinetic.  There is some code here
which will need to go away, but until I have a better handle on how we will
approach this, I will leave it as is.

I do not quite like the interface for update_from_xml and new_from_xml, but
I will think more about this later.

Crud.  Just realized I forgot the write_xml method.  Ah well, it should be a
piece of cake to add.