This commit is mostly a major revision of variable naming conventions inside the store classes.  Search.pm had the column name misnamed as "attr", thus implying we were working with an object attribute when, in fact, we were working with a column name.  Also, cleaned up the documentation for the Search.pm constructor to point to the setter/getters for attributes to ensure that it is clear what args can be passed to the constructor.
Fixed the broken PG.pm test and in the process, cleaned up how some of the XML
tests were working.

Ensured that object attributes were always referred to as "attribute" and
database columns were always referred to as "column".  Also cleaned up a few
misnamed variables that did not properly convey what they were doing.  That was
harder that it sounds.

I did not implement compile-time interfaces as it turned out to be more
difficult than I thought.  The class which implements this exports an import()
function.  So does Exporter::Tidy.  This causes conflicts, but I have some
ideas on how to get around it, but it seemed (for the time being) to be a
low-value feature, so I shelved it.