Somewhat smaller commit today as I stumbled across quite a few tricky issues.
In the XML class, I could not set attributes with mutators as attributes which
contained objects (State, DateTime, etc.) were throwing "Wide character in 
print" errors, so I was forced to fall back on hash access.  However, I have
restructured the code to make the change painless once this issue is resolved.

It would be helpful if $attr->raw() would return the contained object ID instead
of the contained object.  The contained object is trivial to obtain, but if I
am dealing with raw() values, I am exclusively working at the data store level
and I need the ID and not the object.  If this change were in effect, another
special case would go away from the DB.pm code and another method would be
eliminated.


In the XML code, I have to special case the setting of GUID from an object
that is newly instantiated because setting a GUID directly is not allowed.

In TEST::Class::Kinetic, I have added the any_supported(@features) method.  This
makes it easy for classes to test which features are available and skip testing
themselves if those features are not found.  Because of this, if dev_tests are
not enabled, the tests still pass.

get_column() has now been eliminated from the Store::DB class.

Converted $object->{id} to $object->id.

Biggest change was adding a "build_order" key to the object metadata.  This
stores the package names in the reverse order they were added to metadata.  We
can now build objects one at a time and contained objects can be added directly
since we now know they are built in the proper sequence.  This eliminates the
nasty double loop in _build_object_from_hashref.  It also uses more memory, but
we are used to that, eh? :)

Also eliminated some spurious "uninitialized" warnings when some tests were
skipped.