The raw() method will now return IDs for referenced objects. This is true only if you load Kinetic::Meta like so:
  use Kinetic::Meta qw/:with_dbstore_api/;

Note: There is no longer a need to pass this label to Kinetic::Meta::Attribute;
Kinetic::Meta will now pass it on to Attribute, Class, and Type. Modified the
Store::DB class to take advantage of this feature, thus eliminating the 
_get_raw_value() method.

Also changed the schema tests to mock the store classes, so that they can
once again always run, not just when building with one store or another.