Moved DB specific test code to TEST::Kinetic::Store::DB and cleaned up the test inheritance tree to match the regular inheritance tree.  The primary base class, TEST::Kinetic::Store now merely tests whether or not the interfaces work (with full integration testing) without bothering to worry about *which* implementation it is using.
TEST::Kinetic::XML now has the store interface mocked up to remove that
dependency.  Also implemented the write_xml function and corrected the 
interface to dump_xml.

pg and sqlite scheme tests needed SQL updates.  This is finished.

XML now passes the guid to the Store constructor.  Once mutators are implemented
for the rest of the raw attributes, this class will be much cleaner.