Converted the Kinetic/Store/DB.pm module to use a proper lexer/parser for parsing out search parameters.  I have removed more code from DB.pm than I added to write the lexer and parser, so I am quite happy with this change. Additionally, the code is easier to read, too.
The lexer needs more docs, but I so far it appears fairly robust and a quick
glance at t/lexer/code.t will show the data structures that need to be 
emitted by each type of search parameter.  It is actually very straight
forward.

The parser is also fairly clean.  It walks through the data structure produced
by the lexer, converts tokens to search objects and validates that all fields
actually exist in the class being searched.

Both the parser and lexer are procedural in nature, as their is no need for
OO in these modules.  Further, to be able to use XML or strings for searching
the store, all one needs to do is create an appropriate Lexer/XML.pm or 
Lexer/String.pm module and as long as the data structures emitted are
identical, the parser should pick it up without any trouble.