The code is currently in a state of heavy flux while we try to allow proper date parsing in string searches. This is necessary if we are going to build a proper REST server implementation (instead of SOAP).  Now, string and code lexer functions have different names to avoid collisions when they are both imported.  Also, Kinetic::DateTime dates now have the time zone added to them, but we need to figure out some difficulties when they deserialize into XML.
Due to some PG tests failing in obscure ways, the parser tests have been
dramatically expanded.  Specifically, we are running identical code and string
tests side by side to ensure that we really do have almost 100% compatability
between the two search methods.

A number of tests are now running with "no_plan" because we are not done with
this part yet.  Sue me.

I have also simplified the "parse" interface to push the successful parse
validation back into the parser.  The parser now throws an exception if a
full parse cannot be completed.  Code using the parser should not be forced
to handle the verification (particularly since I was not doing it).

Added DateTime::Format::Strptime to the prereqs.