Backed out timezone and nanosecond changes.  I will wait until I have a better understanding of how those should actually be used.
Made ->search($store, STRING => $string) work.  Also added constraints to
string searches (just pass a list of name/value pairs) and a starting set
of tests for the searches.

Now that the interface is more stable, added back in the tests for the search
subs that Store.pm exports (NOT, EQ, BETWEEN, etc.).

Temporarily disable full-text searches for STRING searches until the exact
semantics of them can be nailed down.