More work done on the REST interface.  Currently, the output of the test suite is very messy, even though the tests pass.  Tomorrow I want to try and fine-tune some warning suppression in t/parser/db.t and get rid of the TODO tests for the REST interface.
The REST interface is now capable of doing basic searches by class key.
Started adding direct tests for the REST::Dispatch module (I was using indirect
tests earlier).

The XML generation for REST is a bit sloppy right now.  Hopefully that can
be taken care of soon (it is part of the TODO work.)