Once again the test output is very messy and I hope to get back to it. We still do not know what is causing t/store.t to finish if it does cleanup in an END block, but the problem went away when it was converted to a subroutine.  Jesse, the creator of HTTP::Server::Simple was also not able to provide any insight.
Kinetic::Interface::REST now requires the domain and path be explicitly set.
There is some cleanup I need to do, but this means that XML can provide
proper links without having to guess the above values.

We now have comprehensive tests for the K:I:R interface and the tests for 
Kinetic::Interface::REST::Dispatch eliminated a nasty bug that caused
one test to not be properly tested (do not name tests "can").

If REST is called without a resource, it will now return an XML list of 
resources.

Minor doc tweaks and updated Build.pl.