Significantly reworked the Kinetic::XML interace today.  Will probably have to revisit the Kinetic::Meta::XML interface, too.  The Kinetic::XML interface now uses <attr/> tags for attributes instead of arbitrary tags that are based on the attribute names.  However, when creating an object from XML, the old format is still supported.
Kinetic::XML will now provide a stylesheet link if such is requested in the
constructor.

The REST interface will now let the user browse all the way down to the 
individual instances, but I managed to break fetching the instance.  I will
have to find out how I did that tomorrow.  I will also be adding a new XSLT
stylesheet to display instances (a stub has been added, but it currently
is not working).

Also, XML::SemanticDiff is now throwing all sorts of warnings from is_xml()
in TEST::Kinetic::XML.  This appears to be because namespace URIs are not 
defined in the XML.  This looks like a bug in XML::SemanticDiff
because "undef eq undef" really should not be throwing a warning.