Beefed up XSLT tests in the REST dispatch class.  Also ensured that all HTML generated by XSLT is valid transitional XHTML and soon should be valid strict XHTML.
Moved much of the XSLT handling directly into the REST interface instead of
the dispatch class.  The dispatch class is reponsible for dispatching requests
and answering them.  The REST class provides support information for it.  I
may have some design issues here that should be revisited.

Fixed a bug in the store class which only allowed a single order_by and
sort_order parameter on string searches.

Eliminated the // problem on REST URLs when an empty path was specified.

Removed a few unused classes in some modules.

Added a couple of new exception messages.

Created Kinetic::Util::Constants.  Constants in this module are grouped by
category (currently, "http" and "data_store").  This allows me to pull out
some hard-coded data that was getting strewn around in a bunch of modules.