Updated to latest version of Test::XML to ensure we have the new "is_well_formed_xml" test.
Moved the Kinetic::XML stylesheet url handling to its own method so we can
better control and test it.

Added ATTR_DELIMITER (attribute delimiter) to Kinetic::Util::Constants and
pulled this out of most of the code.

Users can now specify which attributes they want fetched for a class in the
Kinetic::XML class.  Attributes of contained objects also work.  A contained
object is automatically assumed to be a desired attribute to fetch if one of 
its attributes is requested.