This is an "partial" commit of the next phase of the REST interface.  perltidy has been installed and several files have been "tidied" according to the suggestions in Damian books (unlike the naming conventions changes I had been applying yesterday.)
The REST interface has finally been cleaned up and we can once again use the
rest_test.pl program to drill down to instances.  As part of this, the
beginning of proper Web tests has now been implemented.  This will allow us
to drill down through the REST interface as though we were a browser.  This
is the next step prior being able to have full searches available through
the REST interface.

This is only a preliminary commit.  I wanted to get this in while the tests
pass before I break them again while handling the Web testing.