The paging work is mostly finished on the REST side.  I have a few more tests I need to write for it and then it should be good to go.  However, the XSLT is is not yet seeing the pages, so I need to figure out what I did wrong there.
Completed transforming Kinetic::Interface::REST::Dispatch to a class.  This
has tremendously simplified the code internally and also made some of the
hard things easier to manage.

Eliminated an "uninitialized" warning in the DB.pm class.  The warning should
(theoretically) only occur in tests, but since it is harmless, I thought it
was fine to suppress it.