Finally got paging working.  Current page no longer generates a link and all other pages have links such as "[ Page 1 ][ Page 2]", etc.
Backed out plan to convert constants to Readonly.  There appears to be a
problem with regular expressions being assigned to a Readonly value.  I need
to research this.  If I can figure out the problem, we may be able to avoid
constants and allow Readonly values to be interpolated directly.  See
"perldoc Readonly" for discussion of the benefits.

Updated the t/store/rest_test.pl program to create a bunch of "Two" objects
to test paging.  It is pretty slick.  Ugly as sin, but slick.

Also bought "Learning XSLT".  Seems the XSLT Cookbook does not do me a whole
lot of good if I do not actually know XSLT.  That slowed me down quite a bit
today.