Renamed the "_" function in the parser to "match".
Wrote Kinetic::Store::Parser::DB::Overview.  This document contains a
complete BNF grammar (hey, those are hard to write) for our searches.  It
then explains how lexing and parsing work and the intermediate representation
generated by the "parse" function.  I copied some of the documentation from
the Parser::DB class for the end of this.