Added documentation for Kinetic::HOP::Stream and Kinetic::HOP::Lexer.  Tried to ensure I only used one space after periods and I think I have the textwidth=78 issue fixed.
Removed functions from those classes which are unused and are unlikely to be
used and move the iterator_to_stream() function from the lexer to the stream
package.  Also normalized argument order in lexer tokens because Dominus had
them reversed the lexer tranformation anonymous subroutine in the make_lexer
sub. See the docs for details.

Added tests for all public functions in these packages. The tests are not
terribly comprehensive, but they are a decent start and show how to use the
functions if the documentation is not clear.

As for the functions I have removed, I will issue a "mea culpa" if we find
we need to calculate the Hamming function or infinite lists of primes.