Initial tests and documentation for the HOP parser.
This is not done, but there appears to be a serious bug with the way the HOP
parser matches at the end of a stream.  I seem to recall working around some
of the bugs this has caused in the past, but now that I have the start of
some serious tests for it, I have a better idea of what is going on.

As the Kinetic::Store::Parser currently works, I am almost tempted to ignore
the bug and move on, but since I know it is there, I know I have to fix it
because in the future, no one else will have the ability to work on the parser
unless they know about the bug and how to work around it.  This is bad.

For the record, I *suspect* the bug is caused by the tail node of the stream
not being embedded in an array reference.  I could, of course, be smoking
crack.  Speaking of which, after this commit, I will get cracking on that bug.

I do not wish to try and work on the bug prior to this commit as there is a
good chance that fixing the bug will break the K:S:Parser.