Major overhaul of the parsing subsystem.
After resolving the bug in Kinetic::HOP::Parser where concatenate() was
failing at the end of a stream, tests were fleshed out and several new parsers
were built to make the Kinetic::Store::Parser easier to read and maintain.

Kinetic::Store::Parser received the bulk of the attention.  Unfortunately,
working on the parser has been difficult due to inconsistently applied 
eta-conversion (explained in code comments).  By resolving this and applying
the new subparsers added to K:H:Parser, I have managed to reorder the grammar
rules so they can be read top from bottom, just as with a regular grammar.

Further, most of the rules have been considerably simplified and are thus
easier to read.  In doing this, I discovered the BNF grammar in 
Kinetic::Store::Parser::Overview had some errors and fixed them.  I have also
copied the rules from the overview doc and insterted them in front of each
K:S:Parser rule as a comment.  This should also aid future maintenance.

And I would like to take a moment to point out that the parsing system is
effectively a form of declarative programming  (a.k.a., logic programming) :)