I was going to start implementing REST searches today, but I realized I had a few final cleanup tasks to attend to.  They took long enough that I decided to finish the "documentation" week up by cleaning up a few long-standing nits prior to moving on to REST searches next week.
Many packages which blindly imported all functions from another package now
import only those functions which they only use.  This reduces namespace
pollution and makes it easier to determine where those functions came from.

Added the "optional" parser to Kinetic::HOP::Parser.  This was necessary
because the original HOP parser had no concept of 0 or 1.  I could either
build that manually every time I needed it or use the star() parser. 
Unfortuanately, I was using the latter which would have allowed things such
as "name => BETWEEN BETWEEN BETWEEN BETWEEN [ $value1, $value2 ]"  While it
is unlikely anyone is going to do that, there were a few other cases where
this occurred and thus the "official" grammar did not match the behavior of
the parser.

I have also finished adding the error reporting the parser.  It is not
perfect, but if you understand the grammar, you can get a good idea of what
is going on.  Because of the way streams are implemented, it does not do
a great job of reporting context, but it is much better than it was.

Reformatted the XML class as I will start working with this again next week.