Search metadata information has been added to REST XML.  This will allow the XSLT to read this data and create search forms based upon it.  As it turns out, because there is a level of indirection in running through a test Web server, this is really hard to debug (I feel like I have been saying that a lot lately).  Also, any minor change breaks a bunch of XML tests, so I may have some issues with how my tests are set up for REST.
After running testcover, I discovered that coverage for many modules is
abysmal, so I started to take care of that.  So far, that is only one stub
test for Kinetic::Store::Search, but hopefully more will happen later.

Removed a fair amount of duplicate code in the REST dispatch code and made
it somewhat easier to follow.  However, the method for setting args is still
a bit crufty.  I think I may have a design flaw there but I am not sure.

Tried to use Readonly in Kinetic::Util::Constants but found out that the
readonly variables are being assigned to after the import, so their data is
not being exported.  Thus, all Readonly values are undefined outside of that
package.  I will have to see what I can do to replicate that.  It is an
annoying timing issue.