Cleaning up the XSLT.
While spending a day or two really learning XSLT, I was not planning on doing
much with the XSLT we currently have.  However, as I have been learning, I
have discovered a few areas where the current XSLT can be greatly improved and
I have applied those improvements as I have come across them.  It is not a lot
of change, but it definitely helps.  We no longer have an empty <p></p> pair
when page links are not provided and attributes are handled much more
gracefully.

Further improvements will be applied as learned.  Right now, it looks like I
have locked the XSLT into a moderately inflexible style which is dependent on
the overall structure of the XML document rather than the relative structure.
As I learn more, I will break this bad dependency.