Made basic searches work.  Many things have changed today.
Parser has been simplified to make the "fat comma" optional in searches:

  name => EQ thing
  name    EQ thing # same thing

The rest server has a simple hack to allow the serving of static content.
Right now it only recognizes javascript and CSS, but eventually I want it to
recognize a failed request and attempt to redispatch it statically.  If I can
do that, I do not have to do any extra work to add new static document types.

Searches now work.  Run the rest server and type 

  name => "foo"


in the search box.  That shows basic search capability.  You can also override
the port number on the rest server with the --port switch and pass "-i 20" to
limit the number of test objects created.  This allows the server to start up
much faster.

The javascript which controls the search is now served statically.  I might
have the javascript cached for better performance, but it is just a test server
so I am not too worried about it.  I will try to serve the XSLT statically if
I can.

Added 404 error code to http status codes in the constants module.