The full test suite passes again.
Much duplicate code has been removed and factored into the "traits" classes.
I tried to use Class::Trait again and I managed to get it working at one 
point but it turns out that there is a final outstanding issue with the class
which is preventing it from functioning properly.  I have contacted the
author and made some suggestions.  As it turns out, I was bit today by the
fact that I was using exporting and not traits because I managed to override
an existing method with an imported one.  Traits should prevent that.

HTML forms are now built dynamically in tests so in the future, there should be
a lot less cutting and pasting in the test suite.  There are still a few other
areas I could clean things up, but for now, I am just happy that it runs.

Fixed a weird bug with Kinetic::Meta where it was failing on the importing of
the with_dbstore_api symbol.  I do not know why it was working before.