More updates to the search.
I had to stop working on this as I have been working on it too long today. My
eyes are going buggy.

There are some failing tests as this is in transition, but if you run
examples/rest_test.pl, you will get a pretty good idea of where I am going
with this.

Currently, you will notice that "search order" is doing some strange things
because I am in the middle of making it persistent.  Handling this correctly
will allow me to start building other search widgets correctly.  It should
also allow me to get "sortable columns" by merely clicking on their headers.


It is a lot closer but not quite there.