The REST interface is solid again.
Made "Sort Order" sticky and am now building the select widget dynamically.
This should allow future widgets of different types to be easily built as I
have the XSLT in place for it.

Changed "REST" xslt type to "search" to make it match the other xslt names.

I have a bit of a weird design issue.  I have noticed that the bugs I am
currently dealing with have to do with the "constraints" of limit, order by
and sort order.  These constraints, since they are not naturally part of the
Kinetic::Meta package, are what a lot of my code must work around.  I am not
sure what the best way to handle this is, but I have a lot of special case
handling of these.  This is a design flaw and I have to think about it.

I am also having more namespace issues with XML::Genx.  I cannot figure out
how to get it to insert proper namespaces so I gave up.  I will have to 
research this more but I have run into this before.  Also, as I write more
XML manually to get around this, I find that more frequently hardcoding XML
bits.  Again, we have a code smell here and if I do not find a way to tackle
these two issues (the constraints are the larger problem, I think), then this
is going to come back and bite us.