Stabilizing interface prior to XML conversion in REST dispatch class.
1.  Added "type" to search form to allow type=html to be persistent across
    the site navigation.

2.  Switched to tested JS class.

3.  Made the search_data_xml() and search_form() interfaces in the XML and
    HTML traits identical.  This makes building new tests easier.

4.  Eliminated bogus "where is the rest of this test" diag.  The test was for
    the store search but that is a very well-tested method.

I still have some minor problems with not being able to always access the HTML
DOM while using client-side XSLT transformation.  This is only a minor
annoyance, but it is an annoyance nonetheless.  However, the conditions under
which this manifests appear to be unusual enough that it may not be a problem.

5.  (almost forgot) Removed the old search.js since we no longer use it.

Also, at the top of the Kinetic/Search.js class, you will notice a bar
doSearch() function.  For some reason, when I tried to make that a class
method I kept getting "Kinetic.Search.doSearch() is not a function" error
messages in the JS console.  This is a bit ugly, but I am not sure what to do
about this.