Javascript no longer required.
Now, if we turn off Javascript or if the client doest support it, the REST
search will become a query string search.  To this end Ive set up all params
to start with leading underscores if they are "internal" params.  Not sure
if this is the best way, but it works for now.

Fixed long-standing bug whereby forms would not always submit if they were
created with client-side XSLT.  Seems that if we use client-side XSLT processing,
the user-agent doesnt follow the "default action is the last resource"
behavior.

Internal form parameters pulled to constants.

Regretting my decision to embed "STRING" into the uri when we have searches.
Thats the first argument to the search if were using a string search, but
I have found that I need to special case that in a few places.

Updated the javascript tests to handle the new form structure.

Next up is conversting the REST dispatch class to XML::Genx.