Started refactoring of REST dispatch.
I must say that this is frustrating.  I am taking code that was arguably bad
and Im turning it into Java but I really cant think of a better way of
refactoring this.  On one hand, we run the risk of producing incorrect XML.
With the current approach, we are churning out what is guaranteed to be
syntactically correct XML and were writing rather unmaintanable code to do
so.

Well, thats not quite true.  The code will be maintainable in that someone
can read through what Im doing and copy the structure.  As we go further
into this, I might be able to refactor this even more to make something that
is *easy* to write, but for the time being, its slow going.