Heavy refactoring of XML generation in REST dispatch class.
I can now go two ways with this.  I can pull all of this code out into another
module which will generate *Kinetic specific* XML (not be a general purpose
too) or I can leave it where it is and only refactor further if it turns out
to be necessary. Im opting for the latter for the time being.

Aside from adding a lot of documentation, Ive also created and XML
"autohandler" which allows me to specify wrapper XML and pass in the code
to build the core XML.  This should make it easier to produce different
types of XML docs.

Its still not as efficient as it could be, but its cleaner.  Also, I do not
know how maintainable it is.  Of course, that could simply be a function of
XML::Genx.  I think I did the best I could for the time being.

Tomorrow Il get back to work building a richer client.