You can now sort objects by clicking on the column header.
This will do several things.  First, it will persist the current search.
Second, the search form will still be filled in.  Third, if were already
sorting on an attribute and at attribute is sorted in ascending order, the
column header will sort it in a descending order.  Fourth, it breaks the
tests.

With any luck, item number 4 will be fixed early tomorrow.

And figuring out the XSLT for this was a pain in the @$$.  Im such an XSLT
newbie :)