Finally got the tests working for sorting.
I finally figured out a "cleaner" interface for setting default arguments
for REST searches.  As a happy accident, this has the side-effect of not
allowing unknown parameters through.  I am routinely finding out that my
minimalist APIs are too limiting for the rich data we are trying to present.

Cleaned up the CSS a bit.

Pulled the "desired_attributes" code into its own method as it is now being
called for both the instance table and the sort generation.  This has also
made some of the REST dispatch methods shorter.  Many of those methods were
not very cohesive.  I need to watch that.

As a side note, much of the work in the dispatch class is being duplicated in
traits by different algorithms.  This ensures that were not just accidentally
testing the production code by using production code to generate test data.

Note that the new Array::AsHash class must be installed for this code to work.