In REST searches, "Order by" is now a drop-down.
Dispatch class now throws an exception if an unknown class key is set.

search_form and search_data_xml methods in test traits now accept
Array::AsHash arguments.  I think thats just about the last of the methods
which require richer information intests to build the appropriate results.

The next step is going to be adding some "fluff" information to the REST html
output to make it easier to see where you are (right now its not immediately
clear which class you are searching on).  After that I will be breaking out 
individual attributes in the search fields.