Decoupling REST and Store interfaces.
This was a long, hard slog.  I realized that could I not easily break out the
search by attribute because of the way I had hard-coded the constraint names.
This led to a painful renaming of them.  This led to the ugly discovery that
I had tightly coupled the REST and Store interfaces.  I needed to loosely
couple them.  Now, the REST parameters and the arguments to the Store have
a small "translation" layer (a method in the dispatch class, actually) which
ensures that we can rename either side and just update the traslation.

I got this to work after a few hours of pain.  Now Ive thought of a completely
different way I could have tackled it :(  I have no idea which is more correct,
but the other idea would have been simpler to implement.