Guess what?  Ovid doesnt know Jack.  Or JavaScript.
Today was a mainly spent toying with form data to try and figure out the best
way of handling better string searches.  While the code is not yet in place, I
do have a stub javascript method which passes a basic test of building a search
string from multiple form elements.  Next I plan to exand this to allow the 
comparison drop-downs, solidify the tests for it and push this back into the
form generation.  This is going to force me to rework the XSLT a bit, but it
should not be too bad.