Much improved search handling in HTML forms.
I have the basic Javascript worked out to handle search forms broken out
by object properties.  However, this is going to break the heck out of a lot
of tests once I start actually integrating this.  Thus, before I have the
tests broken for an unknown period of time, I want to get this test in place.

What the new Javascript does is allow me to have a form which lookes like
this:

  Name:  [is|is not] [equal to|greater than, etc.] [some value]
  Age:   [is|is not] [equal to|greater than, etc.] [some value]
  Rank:  [is|is not] [equal to|greater than, etc.] [some value]

Where the first two bracketed items are select lists and the last value is a
user entered value.