The search interface now allows much richer searches.
I have broken the Javascript tests.  I have broken the Perl tests.  I have
lost persistence in the search interface in the HMTL.  How much damage can I
do in one day?

The persistence problem raises an interesting problem.  I we go ahead with the
plan to separate out the the object attributes and search by them individually,
we need to have some way of capturing all of the data in the form.  There are
two basic strategies, neither of them are great, but we are trying to do 
something very complicated here.  The first strategy that comes to mind is to 
embed this information in the URL *again* as an easy to parse path info bit
which lets us extract everything.  This is duplicating the information but
it would be the easiest way to handle this.

The next strategy would be to embed this information in a query string.  Same
problem as above but now the URLs get even uglier.

The third strategy, much cleaner but which takes more time to implement is to
create a new type of search.  A "lexer" search would allow the REST dispatch
class to lex the code, extract the information from the lexer stream and persist
the search form with that.  This would possibly require that we create a new
parser which creates an IR for the REST interface.  Another though (one which
just occurred to me) would be to see if there was some way we could have each
objects created for each seach.  If we do that, and I cant see a reason why
not, we should be able to get persistence much easier than I thought.