Searches are now (mostly) persistent in new search interface.
The search interface is going to require a lot of work, but this is probably
*the* key feature that we need to get perfect, so this is OK.

The tests are still broken.  The reason for this is that the tests are just
text.  Theyre too abstract and I cant see what the REST interface is
going to look like just by glancing at the XML.  Thus, when I have a huge
change like what I am currently implementing, the tests are lagging behind.  I
have to constantly check the REST interface to ensure that we are really getting
what we want.  Once the bulk of transition for any phase is working, I go back
and fill in the tests.  I am not quite happy with this, but I cannot see a
better way.

Things accomplished:

* Kinetic::Store::Search now has a few tests.  It also has some extra methods
which allow the IR to provide more useful information about how a search was
initiated.

* A distinction has been made in the documentation between simple and complex
searches.  This is necessary because our "simple" search interface (which is
far more complex than most) is still just a compound search (in a grammatical 
sense).  "Complex" searches are what we would call a "complex" or "compound-complex"
sentence.  These searches are actually very difficult to represent in our
search interface so there will probably have to be two search interfaces.  First
will be the simple search interface and the next will be an "advanced" search
interface.  The latter will just be a text area that allows people to type in
search requests directly.

* Kinetic::Util::Iterator can now return the IR for the request which generated
it.  This is useful if something requests a search but does not know how the
search was generated.  Currently this only completely supports "simple"

searches.

The parser has been updated to allow the following:

  age BETWEEN (21, 42)

Previously that required square brackets.  Its tough to explain, but this
temporarily solves some tricky problems with how people can search.  Also,
research on this functionality revealed a few subtle documentation problems for
searches.  Those problems have been fixed.

* Kinetic::Interface::REST has a bug where I was forgetting to unescape some
data which were being sent through.  This means that building new features by
looking at the resulting Web page means that I might fix bugs but not be able
to conveniently update the tests at the same time.  We can discuss this later
if this is not clear.